HADOOP-17244: resync with hadoop-trunk
catch up with changes in hadoop-trunk test code; moving
the classes to a (hopefully final) destination as standalone
packages
